requirements.txt
aiohttp
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.4
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.18
ci-info==0.3.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
dataclasses-json==0.6.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
etelemetry==0.3.1
executing==2.2.0
fastapi==0.103.2
filelock==3.12.4
fitz==0.0.1.dev2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.38.0
googleapis-common-protos
greenlet==3.0.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.29.5
ipython==8.34.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.2
jiter==0.9.0
jpype1==1.5.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==32.0.1
langchain
langchain-community
langchain-core
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
onnxruntime>=1.17.0
openai==1.66.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.4.0
packaging==23.2
pandas==2.1.1
parso==0.8.4
pathlib==1.0.1
Pillow==10.0.1
platformdirs==4.3.6
posthog==3.0.2
prompt_toolkit==3.0.50
protobuf
prov==2.0.1
psutil==7.0.0
pulsar-client>=3.4.0
pure_eval==0.2.3
pyarrow==13.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.16.1
pyparsing==3.1.2
pypdf==3.16.2
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==309
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.2.1
rdflib==6.3.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
spyder-kernels==3.0.3
SQLAlchemy==2.0.21
stack-data==0.6.3
starlette==0.27.0
streamlit
sympy==1.12
tabula==1.0.5
tabula-py==2.9.3
tenacity==8.2.3
tiktoken==0.9.0
tokenizers==0.14.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.3
traits==6.3.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.17.2
XlsxWriter==3.2.0
yarl==1.9.2
zipp==3.17.0
zstandard==0.23.0
onnxruntime==1.17.0
openpyxl
llama-index==0.9.48
flask
flask_cors
llama-index-readers-llama-parse
pdfplumber